Fix exploded war deployment in application container
I0a7ade3c refactored GWT UI serving and considered two major cases:

1. war deployment - prod mode
2. exploded war deployment - dev mode

Unfortunately one use case was missed:

3. exploded war deployment in application container - prod mode

This broke the serving of GWT UI, because in this mode, the launcher
is loaded from exploded war archive, and thus from the path like:

  <path/to/container>/gerrit/WEB-INF/lib/gerrit-launcher-launcher.jar

This path is erroneously detected as war archive and gerrit_ui path
is tried to serve from this JAR, that obviously can't work.

Apply some heuristic on returned path from gerrit launcher source and
reshuffle the code in static servlet to make it work again.

Similar fix was applied to serving polygerrit UI. The simple case
was also fixed, however, it's still broken because polygerrit cannot
handle correctly context path for now. First request is correctly
served for host:port/gerrit/, but subsequent requests are failing
because context "gerrit" is lost: host:port/styles/main.css and not:
host:port/gerrit/styles/main.css

Test Plan:

Scenario I: Prod mode in embedded Jetty

1. deploy gerrit.war in embedded Jetty
2. confirm that it works as expected

Scenario II: Prod mode in standalone application container

1. setup application container (e.g. Tomcat)
2. deploy gerrit in exploded war mode
3. confirm that it works as expected

Scenario III: Development mode in Eclipse

1. start eclipse SDM debug session
2. confirm that it works as expected

Scenario IV: Development mode polygerrit

1. start polygerrit in development mode from the CLI
2. confirm that it works as expected

Bug: Issue 3758

+++ gerrit-httpd/src/main/java/com/google/gerrit/httpd/raw/DirectoryGwtUiServlet.java
@@ -22,19 +22,23 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
 
-class DeveloperGwtUiServlet extends ResourceServlet {
+class DirectoryGwtUiServlet extends ResourceServlet {
   private static final long serialVersionUID = 1L;
 
   private static final FileTime NOW = FileTime.fromMillis(TimeUtil.nowMs());
 
   private final Path ui;
 
-  DeveloperGwtUiServlet(Cache<Path, Resource> cache, Path unpackedWar)
-      throws IOException {
+  DirectoryGwtUiServlet(Cache<Path, Resource> cache, Path unpackedWar,
+      boolean dev) throws IOException {
     super(cache, false);
     ui = unpackedWar.resolve("gerrit_ui");
-    Files.createDirectory(ui);
-    ui.toFile().deleteOnExit();
+    if (!Files.exists(ui)) {
+      Files.createDirectory(ui);
+    }
+    if (dev) {
+      ui.toFile().deleteOnExit();
+    }
   }
 
   @Override
